geo_point mapping for 5.0
